requirement.txt
aniso8601==8.0.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
cycler==0.10.0
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.10.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
dnspython==2.0.0
email-validator==1.1.2
filelock==3.0.12
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mccabe==0.6.1
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
Pillow==8.0.1
pipenv==2020.11.4
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
twilio==6.48.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
